+++ elmyra.ip.access.epo/elmyra/ip/access/epo/services.py
@@ -1,17 +1,23 @@
 # -*- coding: utf-8 -*-
 # (c) 2013 Andreas Motl, Elmyra UG
-import logging
 import requests
-from pprint import pprint
+import logging
 from pyramid.threadlocal import get_current_request
 from cornice import Service
 from cornice.util import json_error
 from simplejson.scanner import JSONDecodeError
 from elmyra.ip.access.epo.util import object_attributes_to_dict
 from elmyra.ip.access.epo.imageutil import tiff_to_png
+from elmyra.ip.access.epo.client import oauth_client_create
 
 log = logging.getLogger(__name__)
 
+ops_client = None
+def get_ops_client():
+    global ops_client
+    if not ops_client:
+        ops_client = oauth_client_create()
+    return ops_client
 
 # ------------------------------------------
 #   services
@@ -49,9 +55,17 @@ def ops_published_data_search_handler(request):
 def ops_published_data_search(constituents, query, range):
 
     # query EPO OPS REST service
-    url_tpl = "http://ops.epo.org/3.1/rest-services/published-data/search/{constituents}"
+    url_tpl = "https://ops.epo.org/3.1/rest-services/published-data/search/{constituents}"
     url = url_tpl.format(constituents=constituents)
-    response = requests.get(url, headers={'Accept': 'application/json'}, params={'q': query, 'Range': range})
+
+    # v1: anonymous
+    #client = requests
+
+    # v2: with oauth
+    client = get_ops_client()
+
+    response = client.get(url, headers={'Accept': 'application/json'}, params={'q': query, 'Range': range})
+    #print "response:", response
 
     #print "url:", url
     #print "status-code:", response.status_code
@@ -75,14 +89,15 @@ def ops_published_data_search(constituents, query, range):
 @ops_firstdrawing_service.get(renderer='png')
 def ops_firstdrawing_handler(request):
 
-    service_url = 'http://ops.epo.org/3.1/rest-services/'
+    service_url = 'https://ops.epo.org/3.1/rest-services/'
 
+    client = get_ops_client()
 
     # 1. inquire images
-    url_image_inquriy_tpl = 'http://ops.epo.org/3.1/rest-services/published-data/publication/epodoc/{patent}/images'
+    url_image_inquriy_tpl = 'https://ops.epo.org/3.1/rest-services/published-data/publication/epodoc/{patent}/images'
     patent = request.matchdict['patent']
     url_image_inquriy = url_image_inquriy_tpl.format(patent=patent)
-    response = requests.get(url_image_inquriy, headers={'Accept': 'application/json'})
+    response = client.get(url_image_inquriy, headers={'Accept': 'application/json'})
     #print response
     #print dir(response)
     #print response.content
@@ -117,7 +132,7 @@ def ops_firstdrawing_handler(request):
     # http://ops.epo.org/3.1/rest-services/published-data/images/EP/1000000/PA/firstpage.png?Range=1
     # http://ops.epo.org/3.1/rest-services/published-data/images/US/20130311929/A1/thumbnail.tiff?Range=1
     #print "image url:", url
-    response = requests.get(url)
+    response = client.get(url)
     if response.status_code != 200:
         print response
         print response.headers